+++ include/openmc/simulation.h
@@ -87,7 +87,7 @@ void initialize_generation();
 
 //! Full initialization of a particle history
 #pragma omp declare target
-double initialize_history(Particle& p, int index_source);
+double initialize_history(Particle& p, int64_t index_source);
 
 //! Helper function for initialize_history() that is called independently elsewhere
 void initialize_history_partial(Particle& p);
